Ox addition
In many languages it is almost impossible to know if a number has been converted to hexadecimal.  The easiest way to know if a number is in hex is to use the "0x" prefix.  So 0x30 would mean "This is a hexadecimal number whose value is zero(0)."  Using this simple technique means you never again have to wonder if a value that might be, for example, "30" really means "30" or if it means "0".  And the two extra characters are not going to kill your data exchange rate all that much.